+++ app/helpers/application_helper.rb
@@ -1,6 +1,4 @@
 module ApplicationHelper
-  include Pagy::Frontend
-
   def student_group_name(student)
     student.group_name || 'None'
   end